Fixed substitution lacking
Now `unify` type is Term -> Term -> (Valuation,Substitution) -> Maybe (Valuation,Substitution).
This allows to separate substitutions (var,var) from valuations (var,term). Every required
substitution in order to make a term t and the head of a rule r) unify, is then applied to the
goals of the rule. Thus, in any computational step of the proof search, the variables used are
the ones from the goal.